final check before running (minor updates for readability and usability

+++ Install.ps1
@@ -1,38 +1,56 @@
 if (!([Security.Principal.WindowsPrincipal][Security.Principal.WindowsIdentity]::GetCurrent()).IsInRole([Security.Principal.WindowsBuiltInRole] "Administrator")) { 
-    Start-Process powershell.exe "-NoProfile -ExecutionPolicy Bypass -File `"$PSCommandPath`"" -Verb RunAs; exit 
+    Start-Process powershell.exe "-NoProfile -ExecutionPolicy Bypass -File `"$PSCommandPath`"" -Verb RunAs; 
+    exit;
 }
 
 # -----------------------------------------------------------------------------
 # Source script variables
 
-$configureSystemPreferences = $PSScriptRoot + "\configure-system-preferences.ps1"
-$removeUwpApps = $PSScriptRoot + "\remove-uwp-apps.ps1"
-$setUpChoco = $PSScriptRoot + "\setup-choco.ps1"
-$installIss = $PSScriptRoot + "\install-iss.ps1"
-$installDevDependencies = $PSScriptRoot + "\install-dev-dependencies-and-runtimes.ps1"
-$installProductivity = $PSScriptRoot + "\install-productivity.ps1"
-$installEntertainment = $PSScriptRoot + "\install-entertainment.ps1"
-$checkWindowsUpdates = $PSScriptRoot + "\check-windows-updates.ps1"
+$configureSystemPreferences = $PSScriptRoot + "\configure-system-preferences.ps1";
+$removeUwpApps = $PSScriptRoot + "\remove-uwp-apps.ps1";
+$setUpChoco = $PSScriptRoot + "\setup-choco.ps1";
+$installDevDependencies = $PSScriptRoot + "\install-dev-dependencies-and-runtimes.ps1";
+$installProductivity = $PSScriptRoot + "\install-productivity.ps1";
+$installEntertainment = $PSScriptRoot + "\install-entertainment.ps1";
+$checkWindowsUpdates = $PSScriptRoot + "\check-windows-updates.ps1";
 
 # -----------------------------------------------------------------------------
 # Run scripts
 
 # Must always happen first!
-& $setUpChoco
+& $setUpChoco;
 
-& $configureSystemPreferences
-& $removeUwpApps
+& $configureSystemPreferences;
+& $removeUwpApps;
+
+Write-Host "Please review the system configuration and removed rubbish are as expected" -ForegroundColor Yellow;
+$null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
 
 Write-Host ""
-Write-Host "Installing Applications" -ForegroundColor Green
-Write-Host "------------------------------------" -ForegroundColor Green
+Write-Host "Installing Applications" -ForegroundColor Green;
+Write-Host "------------------------------------" -ForegroundColor Green;
 
-& $installIss
-& $installDevDependencies
-& $installProductivity
-& $installEntertainment
+& $installDevDependencies;
+Write-Host "Please review the development dependencies and configuration are as expected" -ForegroundColor Yellow;
+$null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
 
-& $checkWindowsUpdates
+& $installProductivity;
+& $installEntertainment;
+Write-Host "Please review the productivity and entertainment installation is as expected" -ForegroundColor Yellow;
+$null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
 
-# apply updates
-Restart-Computer
+# Updates take a long time, check these last
+& $checkWindowsUpdates;
+
+# -----------------------------------------------------------------------------
+Write-Host "------------------------------------" -ForegroundColor Green;
+Read-Host -Prompt "Setup is done, restart is needed, press [ENTER] to once logs review to (optionally) restart";
+
+Write-Host "Ready to restart and apply changes (Y / N) ?"  $willRestart  -ForegroundColor Yellow;
+
+if ($willRestart -eq "Y") {
+    Write-Host -NoNewLine "Computer will restart on next keystroke";
+    $null = $Host.UI.RawUI.ReadKey('NoEcho,IncludeKeyDown');
+
+    Restart-Computer;
+}
